Add Playwright dependency and update Dockerfile for installation
