+++ Common/package.json
@@ -100,6 +100,7 @@
         "nodemailer": "^6.9.10",
         "otpauth": "^9.3.1",
         "pg": "^8.7.3",
+        "playwright": "^1.48.2",
         "posthog-js": "^1.139.6",
         "prop-types": "^15.8.1",
         "qrcode": "^1.5.3",
