+++ Probe/Utils/Monitors/MonitorTypes/WebsiteMonitor.ts
@@ -29,7 +29,6 @@ export default class WebsiteMonitor {
     url: URL,
     options: {
       retry?: number | undefined;
-      isHeadRequest?: boolean | undefined;
       currentRetryCount?: number | undefined;
       monitorId?: ObjectID | undefined;
       isOnlineCheckRequest?: boolean | undefined;
@@ -45,11 +44,7 @@ export default class WebsiteMonitor {
       options.currentRetryCount = 1;
     }
 
-    let requestType: HTTPMethod = HTTPMethod.GET;
-
-    if (options.isHeadRequest) {
-      requestType = HTTPMethod.HEAD;
-    }
+    const requestType: HTTPMethod = HTTPMethod.GET;
 
     try {
       logger.debug(
@@ -58,25 +53,12 @@ export default class WebsiteMonitor {
         }`,
       );
 
-      let startTime: [number, number] = process.hrtime();
-      let result: WebsiteResponse = await WebsiteRequest.fetch(url, {
-        isHeadRequest: options.isHeadRequest,
+      const startTime: [number, number] = process.hrtime();
+      const result: WebsiteResponse = await WebsiteRequest.fetch(url, {
+        isHeadRequest: false,
         timeout: options.timeout?.toNumber() || 5000,
         doNotFollowRedirects: options.doNotFollowRedirects || false,
       });
-
-      if (
-        result.responseStatusCode >= 400 &&
-        result.responseStatusCode < 600 &&
-        requestType === HTTPMethod.HEAD
-      ) {
-        startTime = process.hrtime();
-        result = await WebsiteRequest.fetch(url, {
-          isHeadRequest: false,
-          timeout: options.timeout?.toNumber() || 5000,
-          doNotFollowRedirects: options.doNotFollowRedirects || false,
-        });
-      }
 
       const endTime: [number, number] = process.hrtime(startTime);
       const responseTimeInMS: PositiveNumber = new PositiveNumber(
