+++ Probe/package.json
@@ -20,11 +20,8 @@
     "@types/ping": "^0.4.4",
     "axios": "^1.7.2",
     "Common": "file:../Common",
-    
-
     "ejs": "^3.1.10",
     "ping": "^0.4.4",
-    "playwright": "1.45.1",
     "ts-node": "^10.9.1"
   },
   "devDependencies": {
